Garbage collect the strdup() replacement for mingw32
Supposedly it was added ten years ago to paper over a mingw32
bug where the returned pointer couldn't be free'd, but I doubt
that the bug still exists assuming the problem was analyzed
correctly in the first place.

If it does, we should notice soon enough as I'd consider
free'ing pointers returned by strdup() a rather common
operation.