Remove an incorrect string termination attempt
It could cause segmentations faults when Privoxy was compiled
without RFC2553 support as *port[NI_MAXSERV - 1] isn't actually
the last position in the allocated memory region.

Given that snprintf() already produces a proper string
the attempt can simply be discarded. The bug was introduced
in 1.105 and is thus not relevant for the ChangeLog.