Add support for external filters
... which allow to process the response body with a script
or program written in any language the platform supports.

External filters are enabled with +external-filter{} after
they have been defined in one of the filter files with
a header line starting with "EXTERNAL-FILTER:".

For this to work, a temporary directory has to be specified
using the newly-added temporary-directory directive.

External filters are experimental and not expected to
work on all platforms (yet).