Introduce the new forwarding type 'forward-webserver'
Currently it is only supported by the forward-override{}
action and there's no config directive with the same
name.

The forwarding type is similar to 'forward', but the
request line only contains the path instead of the
complete URL.

This makes it more convenient to use Privoxy to make
existing websites available as onion services as well.

Many websites serve content with hardcoded URLs and
can't be easily adjusted to change the domain based
on the one used by the client.

Putting Privoxy between Tor and the webserver (or an stunnel
that forwards to the webserver) allows to rewrite headers and
content to make client and server happy at the same time.