Regenerate HTML docs with recent changes
configure.in has been locally reverted to r1.204
to get 3.0.26 stable documentation that can be put
on the webserver before the next release.

Tidied with tidy 5.2.0 from http://www.html-tidy.org/.

As an expected side effect the diff is bloated enough
to be considered a "rewrite" by git (again), but using
a more recent tidy version should make it easier to
prevent frequent diff bloat in the future when building
on different platforms.

As an unexpected side-effect some pre tags are messed
up now and their first line contains bogus white-space
that causes rendering issues. This regression should
be analysed and fixed or worked around before the next
release but doesn't seem critical.

It's conceivable that the tidy upgrade caused less
obvious regressions that have yet to be found.