workaround for tidy -indent -wrap breaking <pre> formatting
tidy -indent -wrap 78  messes up the formatting of a long
line following a <pre> tag by splitting the line into <pre>
and the text and indenting both.  eg
<pre>a really long line</pre>
is turned into
          <pre>
          a really long line</pre>
and we end up with lots of leading spaces.

change the line wrap to column 120 and hope that's enough
to keep this bug from happening