have docbook generated html files be straight ascii
Dealing with a mixture of ISO-8859 and UTF-8 files is.. problematic.
To keep the number of issues down, keep everything in ASCII as much
as possible.

We have two docbook stylesheet sections - 'html' for generated
html files where the output is supposed to be plain 7-bit ASCII
and 'print' for generated text files where the output is
latin1/ISO-8859

So the makefile has two docbook calls
  $(DB)     for generating html ascii output
  $(DB_TXT) for generating latin1 text output (eg AUTHORS)

and two w3m calls
  $(W3M_DUMP)      for reading docbook output and creating ASCII output
  $(W3M_DUMP_UTF8) for reading docbook output and creating UTF-8 output