Add OpenSSL support
To enable it, configure with "--with-openssl".

This is mainly useful for platforms that ship with
OpenSSL in the base so the "special exception" from
section 3 of the GPLv2 applies.

The author (Maxim Antonov) reports about 100% more requests
per second when using OpenSSL instead of MbedTLS which the
committer could confirm. For details see:
https://sourceforge.net/p/ijbswa/patches/143/

This commit has a couple of issues that will be addressed
in follow-up commits.