Add keep-alive support for https-inspected connections
If a connection to the server has been established it
is reused.

If the previous request was blocked and no connection to
the server is open the connection is dropped so the client
can retry on a fresh connection.

Forwarding settings are currently ignored when handling
https-inspected requests on an already established connection.

Sponsored by: Robert Klemme