cgi_edit_actions_submit(): Check the toggle state of filters until no filters are left
Previously we would stop looking after the first filter
index wasn't found in the request URL.

This worked in case of "split-large-forms 0" but resulted in
filter state being ignored in case of "split-large-forms 1"
which leads to request URLs that only contain a subset of
the filters.

Reported by withoutname in #921.