Allow to use wolfSSL for https inspection
It's licensed under GPlv2 or later and unlike mbedTLS
there don't seem to be plans to change the license.

As a bonus, wolfSSL supports TLS 1.3 and can be significantly
faster than mbedTLS. Mainly tested on ElectroBSD amd64 where
it can compete with OpenSSL and LibreSSL:
https://www.fabiankeil.de/gehacktes/privoxy-tls-benchmarks/

To enable the support, install wolfSSL and run ./configure
with the --with-wolfssl option.

Privoxy users and packagers that currently build Privoxy
binaries with mbedTLS may want to consider using wolfSSL
in the future once it has been properly tested.

Sponsored by: Privoxy project funds collected at SPI