remove typo
"_CYGWIN" is not defined for the cygwin or mingw cross compiler:

$ echo | gcc -dM -E -xc - | grep -i cygwin

$ echo | i686-w64-mingw32-gcc -dM -E -xc - | grep -i cygwin

Maybe _CYGWIN is a typo left over from the days when the same gcc
compiler was used to compile native windows (gcc -mno-cygwin) and
cygwin apps?  In any case, !defined(_CYGWIN) is true so removing
the test just makes the code a bit clearer.