Add missing client-body-tagger data to the action_type_info[] struct
... so lookups based on the action index work correctly again.

Prevents assertion failures or segfaults when trying to edit
an action file with the CGI editor.

The type of failure depended on whether or not assertions
were enabled and on whether or not Privoxy had been compiled
with FEATURE_EXTERNAL_FILTERS.

Regression introduced in Privoxy 3.0.34.

Patch submitted by Aaron Li in #940.